requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Bottleneck==1.3.2
certifi==2021.10.8
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/croot-2p6q7e2r/cffi_1642701115063/work
chardet==4.0.0
cryptography @ file:///opt/concourse/worker/volumes/live/732d2752-e4cb-4715-726b-56d890f169fb/volume/cryptography_1639414586414/work
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.2
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
lml==0.1.0
MarkupPy==1.14
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/f196a661-8e33-4aac-63bc-efb2ff50e035/volume/mkl_random_1626186080069/work
mkl-service==2.4.0
numexpr @ file:///opt/concourse/worker/volumes/live/52f18ae0-3912-48f4-7832-7d22abf43616/volume/numexpr_1640704220628/work
numpy @ file:///opt/concourse/worker/volumes/live/ceac9fd4-1ae1-48f8-612a-0c99bf783d0e/volume/numpy_and_numpy_base_1634106709647/work
odfpy==1.4.1
openpyxl==3.0.9
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyexcel==0.6.7
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyMySQL @ file:///opt/concourse/worker/volumes/live/91bc5664-f34a-482d-7efc-f19bbd7a1347/volume/pymysql_1610482895985/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse==0.4.2
tablib==3.2.0
texttable==1.6.4
xlrd==2.0.1
xlwt==1.3.0
